WIP: RAG Chat logic with sessions and LLM for structuring the output.

+++ models/document.go
@@ -3,9 +3,9 @@ package models
 import "time"
 
 type Document struct {
-	ID           int       `json:"id" db:"id"`
-	UserID       int       `json:"-" db:"user_id"`
-	CollectionID int       `json:"collection_id" db:"collection_id"`
+	ID           int64     `json:"id" db:"id"`
+	UserID       int64     `json:"-" db:"user_id"`
+	CollectionID int64     `json:"collection_id" db:"collection_id"`
 	FileName     string    `json:"file_name" db:"file_name"`
 	IsIndexed    bool      `json:"is_indexed" db:"is_indexed"`
 	DateCreated  time.Time `json:"date_created" db:"date_created"`
